Add a special response header to signify that the response is from cache
Currently there is no way for the client to identify whether the response
is from client cache or from server. So in line with what caches such as
varnish do - add a new X-HTTPCache4J header which contains the date time of
when the item was cached.

So a header is added to the response whenever the response is from the
cache.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>