Keep HTTPMethod backward compatible
The 2 most used methods of enum is name() and valueOf(). When converting
enum to class though valueOf was implemented by name() was not and thus
upgrading to 3.2 may cause compile time error, thus added it as deprecated
to keep it compatible and notify the user not to use it as the same time.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>