Switch to using Escenic's HTTP ESI Testbed webapp as an external library
The main reason for this switch is to be harness any changes, fixes made to
that web app rather than maintaining it locally.

Add the snapshot repository where http-esi-testbed is deployed. With recent
changes it supports response date out-of-the-box and also supports basic
and digest authentication.

Fixed some imports and dependencies.

Next step would be to delete the codes that is being used from external
source.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>