Cache Hit occurring when revalidation with backend occurs, reporting incorrect Age header
The HTTP Response is returning a cache hit, and a Age head when content
is initially revalidated from the back end.  For example:

- Initial Request results in "Cache-Control: max-age=10"
- Cache has item added
- 2nd Request, in 1 second's time, hits the cache.  Response has "Age:
1",and cache is hit
- Wait for 12 seconds.
- 3rd Request, the backend is contacted, new content (body is returned)
along with "Cache-Control: max-age-10".  However, response has "Age:
13", and Response.isCached() is true.  
- 4th Request, in one second time, the cache is hit and Age is 1.

"When a response is generated from a cache entry, the cache MUST include
a single Age header field in the response with a value equal to the
cache entry's current_age." rfc2616, 13.2.3 

Given the above, the Age shouldn't really be 13.

So I've changed HTTPCache and HTTPCacheHelper to  addresses the above
scenario,such that the validation with the backend server results in
Reponse.isCached() false and no Age header.